Blog Detail view increment problem

+++ client/src/api/userService.js
@@ -7,13 +7,10 @@ const userService = {
   updateProfile: (userData) => apiClient.put('/users/profile', userData),
   getCurrentUser: () => apiClient.get('/users/profile'),
   deleteAccount: (userId, deleteBlogs, token) =>
-    apiClient.delete(
-      `/users/delete/${userId}`,
-      {
-        data: { deleteBlogs },
-        headers: { Authorization: `Bearer ${token}` }
-      }
-    )
+    apiClient.delete(`/users/delete/${userId}`, {
+      data: { deleteBlogs },
+      headers: { Authorization: `Bearer ${token}` },
+    }),
 };
 
 export default userService;
